requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.1.1
django-ckeditor-5==0.2.13
django-cors-headers==4.4.0
django-filter==24.3
django-import-export==4.1.1
django-modeltranslation==0.19.9
django-parler==2.3
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
pytz==2024.2
PyYAML==6.0.2
setuptools==74.1.2
sqlparse==0.5.1
tablib==3.5.0
tzdata==2024.1
uritemplate==4.1.1
